backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.34.0
python-multipart==0.0.9

# Database — SQLite (local dev)
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Auth — local JWT
python-jose[cryptography]==3.3.0
cryptography==42.0.8
bcrypt==4.1.2

# Config
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.1.0

# Image processing
Pillow==10.4.0

# Excel import
openpyxl==3.1.2

# Google AI — prompt generation + image generation
google-auth>=2.29.0
google-genai>=1.0.0

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
